Migrated controller code
