+++ src/main/kotlin/io/craigmiller160/authmanagementservice/service/ClientRoleService.kt
@@ -33,4 +33,12 @@ class ClientRoleService (
         return RoleList(roles)
     }
 
+    @Transactional
+    fun updateRole(clientId: Long, roleId: Long, role: Role): Role {
+        roleRepo.findByClientIdAndId(clientId, roleId)
+                ?: throw EntityNotFoundException("Role not found for ClientID $clientId and RoleID $roleId")
+        val finalRole = role.copy(id = roleId, clientId = clientId)
+        return roleRepo.save(finalRole)
+    }
+
 }
